clean up swagger specs for UI usage (#43)
Changes to support using the swagger specs to generate API classes in UI

* add not null assertions

* update types to generate proper swagger specs

+++ music-api/src/main/java/music/mapper/TrackMapper.java
@@ -46,7 +46,7 @@ public interface TrackMapper {
 
     void markDeletedById(@Param("id")long id, @Param("deletedInd") boolean deletedInd);
 
-    void setRatingById(long id, byte rating);
+    void setRatingById(long id, int rating);
 
     void updateFieldById(long id, String field, Object newValue, String type);
